+++ pkg/image/definition_test.go
@@ -54,19 +54,22 @@ func TestParse(t *testing.T) {
 	assert.Equal(t, userConfigs[0].PrimaryGroup, "admin")
 	assert.Len(t, userConfigs[0].SecondaryGroups, 1)
 	assert.Equal(t, "wheel", userConfigs[0].SecondaryGroups[0])
+	assert.True(t, userConfigs[0].CreateHome)
 
 	assert.Equal(t, "beta", userConfigs[1].Username)
 	assert.Equal(t, "$6$GHjiVHm2AT.Qxznz$1CwDuEBM1546E/sVE1Gn1y4JoGzW58wrckyx3jj2QnphFmceS6b/qFtkjw1cp7LSJNW1OcLe/EeIxDDHqZU6o1", userConfigs[1].EncryptedPassword)
 	assert.Nil(t, userConfigs[1].SSHKeys)
 	assert.Equal(t, userConfigs[1].PrimaryGroup, "")
 	assert.Len(t, userConfigs[1].SecondaryGroups, 0)
+	assert.False(t, userConfigs[1].CreateHome)
 
 	assert.Equal(t, "gamma", userConfigs[2].Username)
 	assert.Equal(t, "", userConfigs[2].EncryptedPassword)
 	assert.Len(t, userConfigs[2].SSHKeys, 1)
 	assert.Contains(t, userConfigs[2].SSHKeys[0], "ssh-rsa BBBBB3")
-	assert.Equal(t, userConfigs[1].PrimaryGroup, "")
+	assert.Equal(t, userConfigs[2].PrimaryGroup, "")
 	assert.Len(t, userConfigs[2].SecondaryGroups, 0)
+	assert.False(t, userConfigs[2].CreateHome)
 
 	// Operating System -> Systemd
 	systemd := definition.OperatingSystem.Systemd